#70 move mock data to interactor
