+++ lib/features/home_select/presentation/bloc/promo/home_select_promo_bloc.dart
@@ -1,40 +1,42 @@
 import 'dart:async';
 
-import 'package:dogdom/features/home_select/domain/promo_card_model.dart';
+import 'package:dogdom/features/home_select/domain/interactors/promo_interactor.dart';
 import 'package:dogdom/features/home_select/presentation/bloc/promo/home_select_promo_bloc_models.dart';
-import 'package:dogdom/features/home_select/presentation/ui/widgets/promo_card.dart';
-import 'package:dogdom/generated/assets.gen.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:injectable/injectable.dart';
 
 @injectable
 class HomeSelectPagePromoBloc
     extends Bloc<HomeSelectPagePromoEvent, HomeSelectPagePromoState> {
-  HomeSelectPagePromoBloc()
-      : super(
+  @protected
+  final PromoInteractor interactor;
+
+  HomeSelectPagePromoBloc({
+    required this.interactor,
+  }) : super(
           HomeSelectPagePromoState.loading(),
         ) {
     on<HomeSelectPagePromoEventInit>(_init);
     on<HomeSelectPagePromoEventTap>(_onTap);
   }
 
-  Future<FutureOr<void>> _init(
+  Future<void> _init(
     HomeSelectPagePromoEventInit event,
     Emitter<HomeSelectPagePromoState> emit,
   ) async {
     emit(HomeSelectPagePromoState.loading());
-    await Future.delayed(Duration(seconds: 2));
     emit(
       HomeSelectPagePromoState.data(
-        promoCardList: _buildList(),
+        promoCardList: await interactor.getData(),
       ),
     );
   }
 
-  FutureOr<void> _onTap(
+  Future<void> _onTap(
     HomeSelectPagePromoEventTap event,
     Emitter<HomeSelectPagePromoState> emit,
-  ) {
+  ) async {
     final tempDataState = state.data;
     emit(
       HomeSelectPagePromoState.onTap(
@@ -43,33 +45,4 @@ class HomeSelectPagePromoBloc
     );
     emit(tempDataState);
   }
-}
-
-// TODO get data from BLoC
-List<PromoCard> _buildList() {
-  return [
-    PromoCard(
-      model: PromoCardModelTakeHome(
-        title: 'Take me Home',
-        description: 'Buy me a bowl of food.',
-        imgPath: Assets.imageMock.homeSelectCarousel2.path,
-        buttonText: 'Let me',
-      ),
-    ),
-    PromoCard(
-      model: PromoCardModelDonate(
-        title: 'Feed me',
-        description: 'Buy me a bowl of food.',
-        imgPath: Assets.imageMock.homeSelectCarousel1.path,
-      ),
-    ),
-    PromoCard(
-      model: PromoCardModelTakeHome(
-        title: 'Take me Home',
-        description: 'Buy me a bowl of food.',
-        imgPath: Assets.imageMock.homeSelectCarousel2.path,
-        buttonText: 'Let me',
-      ),
-    ),
-  ];
 }
